UI Who/Skill Update
Skills and Who boxes are now functional.

It’s also worth noting that the fake imps from the impme command no
longer are counted. Well that’s not exactly true, they are counted but
they are uncounted. The count is decremented and the time changed back
to its original number when the (but you were mistaken) appears.