Loader change
Did a change to how the loader works. Did away with the /reload, /load,
and /unload aliases. These will be replaced with fixed aliases in the
package. Also, files are found by _core and loaded in rather than their
own separate file. This also means that I can add in files to the
folders without having to hard-code a load in for them.