Updates
Found the proper link to download the development version of the
scripts from github. I also learned that the luazip module is part of
mudlet, so I’m going to start working on using that to unpack the zip
file. Honestly though, I think I could just rename the zip file to
.mpackage and run an uninstall/install and it’d work fine. We’ll see
though.