Sz update
For some reason, I was not grabbing sis with the other trigger so this
one was added. This is fine though, because later on down the line i
want to add a unique trigger for each chat event so that we don’t
accidentally add in random text. Also, being able to grab all of a line
instead of it having split would be nice.